ML_models/requirements.txt
catboost==1.2
hydra-core==1.3.2
numpy==1.24.3
omegaconf==2.3.0
pandas==2.0.1
scikit-learn==1.2.2
xgboost==1.7.5